Comment out asserts that broke distrib support.
Something about the way that we defined the distrib functions made LLVM think that some things weren't defined that actually were.  Commenting out the asserts makes things work again.  Ideally, we'd instead fix roadrunner's declarations of distrib functions, but I can't tell how to do that.  I've filed an issue tracker at the llvm site at https://llvm.discourse.group/t/how-to-tell-the-llvm-6-jit-executionengine-that-a-function-exists/1147 so if that ever gets answered, we could revert this change and fix it in roadrunner instead.