Change exit() to throw().
In particular, when loading a roadrunner file, if the file is old/corrupted, the LLVM would exit entirely, taking out Python with it.  Changing this to a throw means that the user can actually recover.

It's not clear why llvm uses exit() instead of throw() in the first place, though it's clearly a deliberate design decision (there are no throws anywhere in the codebase, and several exit() calls).  https://llvm.discourse.group/t/why-does-llvm-exit-instead-of-throw/1629 to find out why.